+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+==========
+
+* Improve memory usage for windows where `Option<T>` is not optimized
+
 0.1.1
 =====